Split up bluetooth code into its own file
