+++ bluetooth/bluetooth_driver.h
@@ -51,6 +51,13 @@ typedef struct bluetooth_driver
 extern bluetooth_driver_t bluetooth_bluetoothctl;
 extern bluetooth_driver_t bluetooth_bluez;
 
+typedef struct
+{
+   const bluetooth_driver_t *drv;
+   void *data;
+   bool active;
+} bluetooth_driver_state_t;
+
 /**
  * config_get_bluetooth_driver_options:
  *
@@ -75,6 +82,8 @@ bool driver_bluetooth_connect_device(unsigned i);
 bool driver_bluetooth_remove_device(unsigned i);
 
 bool bluetooth_driver_ctl(enum rarch_bluetooth_ctl_state state, void *data);
+
+bluetooth_driver_state_t *bluetooth_state_get_ptr(void);
 
 extern const bluetooth_driver_t *bluetooth_drivers[];
