+++ network/wifi_driver.h
@@ -121,6 +121,8 @@ void driver_wifi_tether_start_stop(bool start, char* configfile);
 
 bool wifi_driver_ctl(enum rarch_wifi_ctl_state state, void *data);
 
+wifi_driver_state_t *wifi_state_get_ptr(void);
+
 extern const wifi_driver_t *wifi_drivers[];
 
 RETRO_END_DECLS
