+++ qb/config.params.sh
@@ -1,6 +1,7 @@
 HAVE_LIBRETRO=             # Libretro library used
 HAVE_ASSETS_DIR=           # Assets install directory
 HAVE_CORE_INFO_CACHE=yes   # Core info cache support
+HAVE_BLUETOOTH=no          # Bluetooth support
 HAVE_NVDA=yes              # NVDA support
 HAVE_PATCH=yes             # Softpatching support (BPS/IPS/UPS)
 HAVE_SAPI=no               # SAPI support
